Add live search for products (using jsroutes, requirejs, coffeescript)
Live search is performed while typing the search query. Several keyup
events are "debounced" (within 300 millis), so that not each keyup
event triggers a search, but only the last in this duration.

The browser history is updated accordingly, and backward/forward navigation
is also supported (using livesearch).

The api returns rendered products (html) wrapped in a json response, so
that the browser can update the related/specified element (by id).